Fix TryGetNextNonSwitchArgumentOrError.
